Extract and fix batch send tests
